Separates common room-placement logic from BRP specific logic
